Generalize crt to Integral and use crt in 2019 day 12 part 2

+++ src/main/scala/eu/sim642/adventofcode2016/Day15.scala
@@ -23,8 +23,10 @@ object Day15 {
 
   object Part1 extends Part {
     override def firstPressTime(disks: Seq[Disk]): Int = {
-      val ans = disks.map({ case Disk(i, posCount, initialPos) => ((-(initialPos + i)) %+ posCount, posCount) })
-      NumberTheory.crt(ans)._1
+      val ans = disks.map({ case Disk(i, posCount, initialPos) =>
+        (((-(initialPos + i)) %+ posCount).toLong, posCount.toLong)
+      })
+      NumberTheory.crt(ans)._1.toInt
     }
   }
